Refactoring stocks provider
