+++ src/shared/container/SocksProvider/IStocksProvider.ts
@@ -1,5 +1,11 @@
-import { IFindStockDTO } from "../../../modules/stocks/dtos/IFindStockDTO";
+import { IFindStockDTO, IGetHistoricDTO } from "../../../modules/stocks/dtos/IStockDTO";
+
+export interface IResponse {
+    name: string,
+    prices: []
+}
 
 export interface IStocksProvider {
-    fetch(stock_name: string): Promise<IFindStockDTO>
+    fetchQuote(stock_name: string): Promise<IFindStockDTO>,
+    fetchHistoric({stock_name, from, to}: IGetHistoricDTO): Promise<IResponse>
 }
